Code cleanup
Remove unused functions and arguments ; Convert array indexes from INT to LONG ; Remove unimportant 'omp simd' pragmas

+++ src/Kernels/flux_loops.h
@@ -40,7 +40,7 @@ void compute_flux_edge(
     #else
         double *restrict fluxes
         #ifdef COLOURED_CONFLICT_AVOIDANCE
-        , int serial_section_start
+        , long serial_section_start
         #endif
     #endif
     );
@@ -59,7 +59,7 @@ void compute_flux_edge_crippled(
     #else
         double *restrict fluxes
         #ifdef COLOURED_CONFLICT_AVOIDANCE
-        , int serial_section_start
+        , long serial_section_start
         #endif
     #endif
     );
