+++ src/Meshing/graph.h
@@ -5,51 +5,52 @@
 
 // Count degree of each point.
 int* count_point_degrees(
-	int npoints, 
+	long npoints, 
 	const edge_neighbour* edges, 
-	int edges_start_idx, 
-	int edges_end_idx);
+	long edges_start_idx, 
+	long edges_end_idx);
 
 // Count internal-degree of each point.
 int* count_point_internal_degrees(
-	int npoints, 
+	long npoints, 
 	const edge_neighbour* edges, 
-	int edges_start_idx, 
-	int edges_end_idx);
+	long edges_start_idx, 
+	long edges_end_idx);
 
 // Identify point-wise internal neighbours
-int** identify_pointwise_internal_neighbours(
+long** identify_pointwise_internal_neighbours(
 	const int* point_degree, 
-	int npoints, 
+	long npoints, 
 	const edge_neighbour* edges, 
-	int edges_start_idx, 
-	int edges_end_idx);
+	long edges_start_idx, 
+	long edges_end_idx);
 
 // Identify edges connected to each point.
-int** construct_point_to_edge_mapping(
+long** construct_point_to_edge_mapping(
 	const int* point_degree, 
-	int npoints, 
+	long npoints, 
 	const edge_neighbour* edges, 
-	int edges_start_idx, 
-	int edges_end_idx);
+	long edges_start_idx, 
+	long edges_end_idx);
 
 // Identify edgewise neighbours
 int* count_edge_degrees(
 	const int* point_degree,
-	int** point_to_edges, 
+	long** point_to_edges, 
 	const edge_neighbour* edges, 
-	int nedges, 
-	int edges_start, 
-	int edges_end);
+	long nedges, 
+	long edges_start, 
+	long edges_end);
 
 // Generate mapping of edge_id -> {edge_ids}
-int** identify_edgewise_neighbours(
-	int** point_to_edges, 
+long** identify_edgewise_neighbours(
+	long** point_to_edges, 
 	const int* point_degree,
 	const int* edge_degree,
 	const edge_neighbour* edges,
-	int nedges,
-	int edges_start_idx,
-	int edges_end_idx);
+	long nedges,
+	long edges_start_idx,
+	long edges_end_idx);
 
 #endif
+
